+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Fixed
+- Raise `ImportError` when initialising `OpenAIEmbedder` without the `openai` package
+  installed.
+
+
 ## [v1.2.0] - 2024-08-15
 ### Added
 - Added an `OpenAIEmbedder` that uses the OpenAI Embeddings API to embed documents.